bug-nottifications

+++ store/AppActions.js
@@ -100,32 +100,59 @@ export const FileSystems = (coordinate) => {
     ReactNativeFs.DocumentDirectoryPath + 'datas.json'
   let readFile;
   return async (dispatch) => {
-    // let fileExist = ReactNativeFs.exists(PATH)
-    // if (fileExist) {
-    //   readFile = await ReactNativeFs.readFile(PATH);
-    //   let parseFile = JSON.parse(readFile);
-    //   console.log(`Parse ${parseFile}`)
-    //   dispatch(ActionCreators.LoadingFile(parseFile))
-    //   if (EqualDate(parseFile.data.current.dt)) {
-    //     dispatch(ActionCreators.LoadingFile(parseFile))
-    //     return;
-    //   }
-      // else {
-      //   try {
-      //     const response = await GetWeather.getYesterdayWeather(coordinate)
-      //      ReactNativeFs.writeFile(PATH, JSON.stringify(response))
-      //     const readFile = await JSON.parse(ReactNativeFs.readFile(PATH))
-      //     console.log(`response ${response}`)
-      //     dispatch(ActionCreators.LoadingFile(response))
-      //   } catch (error) {
-      //     dispatch(ActionCreators
-      //       .ErrorMessage(error))
-      //   }
-      // }
-    // } else {
+    let fileExist = ReactNativeFs.exists(PATH)
+    if (fileExist) {
+      readFile = await ReactNativeFs.readFile(PATH);
+      let parseFile = JSON.parse(readFile);
+      if (EqualDate(parseFile.data.current.dt)) {
+        dispatch(ActionCreators.LoadingFile(parseFile))
+        Notifications.postLocalNotification({
+          body: "File with yesterday forecast succesfully donwloaded",
+          title: "File Downloaded",
+          sound: "chime.aiff",
+          category: "SOME_CATEGORY",
+          link: "localNotificationLink",
+          // fireDate: Date.now() + 150,
+        })
+        return;
+      }
+      else {
+        try {
+          const response = await GetWeather.getYesterdayWeather(coordinate)
+          ReactNativeFs.writeFile(PATH, JSON.stringify(response)).then(() => {
+            console.log('Notifications')
+            Notifications.postLocalNotification({
+              body: "File with yesterday forecast succesfully donwloaded",
+              title: "File Downloaded",
+              sound: "chime.aiff",
+              category: "SOME_CATEGORY",
+              link: "localNotificationLink",
+              // fireDate: Date.now() + 150,
+            })
+          })
+          const readFile = await JSON.parse(ReactNativeFs.readFile(PATH))
+          console.log(`response ${response}`)
+          dispatch(ActionCreators.LoadingFile(response))
+        } catch (error) {
+          dispatch(ActionCreators
+            .ErrorMessage(error))
+        }
+      }
+    } else {
       try {
         const response = await GetWeather.getYesterdayWeather(coordinate)
         ReactNativeFs.writeFile(PATH, JSON.stringify(response))
+          .then(() => {
+            console.log('Notifications')
+            Notifications.postLocalNotification({
+              body: "File with yesterday forecast succesfully donwloaded",
+              title: "File Downloaded",
+              sound: "chime.aiff",
+              category: "SOME_CATEGORY",
+              link: "localNotificationLink",
+              // fireDate: Date.now() + 150,
+            })
+          })
         const redFile = await ReactNativeFs.readFile(PATH)
         const parseFile = await JSON.parse(redFile)
         dispatch(ActionCreators.LoadingFile(response))
@@ -134,7 +161,7 @@ export const FileSystems = (coordinate) => {
       }
     }
   }
-// }
+}
 
 
 
@@ -143,19 +170,18 @@ export const OpenFile = () => {
     var ReactNativeFs = require('react-native-fs');
     var path = Platform.OS.toLowerCase() === 'android' ? ReactNativeFs.DocumentDirectoryPath + 'datas.txt' : ReactNativeFs.DocumentDirectoryPath + 'datas.json';
     let exists = await ReactNativeFs.exists(path);
-    console.log("OpenFile")
     if (exists) {
-      //   Notifications.postLocalNotification({
-      //   body: "File with yesterday forecast succesfully donwloaded",
-      //   title: "File Downloaded",
-      //   sound: "chime.aiff",
-      //   category: "SOME_CATEGORY",
-      //   link: "localNotificationLink",
-      //   // fireDate: Date.now() + 150,
-      // })
+
       FileViewer.open(path, { showOpenWithDialog: true, showAppsSuggestions: true })
         .then(() => {
-
+          Notifications.postLocalNotification({
+            body: "File with yesterday forecast succesfully donwloaded",
+            title: "File Downloaded",
+            sound: "chime.aiff",
+            category: "SOME_CATEGORY",
+            link: "localNotificationLink",
+            // fireDate: Date.now() + 150,
+          })
         })
     }
   }
